% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{build.labels}
\alias{build.labels}
\title{Builds labels for plots}
\usage{
build.labels(input, order.by, name, digits = 2)
}
\arguments{
\item{input}{table}

\item{order.by}{what should be ordered (to show max, min and current)}

\item{name}{name of grouping variable}

\item{digits}{digits on numbers}
}
\value{

}
\description{
Builds labels for plots
}
