% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid19.pt}
\alias{covid19.pt}
\title{PT DGS and EU CDC data combined}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 47 rows and 10 columns.
}
\usage{
covid19.pt
}
\description{
PT DGS and EU CDC data combined
}
\examples{
head(covid19.pt)
}
\keyword{data}
