% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dgs.pt}
\alias{dgs.pt}
\title{PT DGS data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 5 columns.
}
\usage{
dgs.pt
}
\description{
PT DGS data
}
\examples{
head(dgs.pt)
}
\keyword{data}
