% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapper.R
\name{download.report}
\alias{download.report}
\title{Download report}
\usage{
download.report(only.date = NULL, index = 1)
}
\arguments{
\item{only.date}{filter out if date isn't the same (NULL for keep whatever report it is)}

\item{index}{index of report to extract (1 is the latest)}
}
\value{
report contents
}
\description{
Download report
}
