% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapper.R
\name{download.updated.pt}
\alias{download.updated.pt}
\title{Merge EU CDC and PT DGS data (all in one)}
\usage{
download.updated.pt()
}
\value{
updated data
}
\description{
Merge EU CDC and PT DGS data (all in one)
}
