% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapper.R
\name{download_all_reports}
\alias{download_all_reports}
\title{Downloads all reports that are not in cache}
\usage{
download_all_reports()
}
\value{
data from dgs
}
\description{
Downloads all reports that are not in cache
}
