% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapper.R
\name{extract_cases}
\alias{extract_cases}
\title{Extract number of cases from report}
\usage{
extract_cases(page)
}
\arguments{
\item{page}{page strings from document}
}
\value{
a number
}
\description{
Extract number of cases from report
}
