% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapper.R
\name{extract_cases2}
\alias{extract_cases2}
\title{Extract number of cases from report (from 2020/08/17)}
\usage{
extract_cases2(page)
}
\arguments{
\item{page}{page strings from document}
}
\value{
a number
}
\description{
Extract number of cases from report (from 2020/08/17)
}
