% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_scrapper.R
\name{extract_deaths}
\alias{extract_deaths}
\title{Extract number of deaths from report}
\usage{
extract_deaths(page)
}
\arguments{
\item{page}{page strings from document}
}
\value{
a number
}
\description{
Extract number of deaths from report
}
