% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_scrapper.R
\name{extract_generic.two}
\alias{extract_generic.two}
\title{Generic extractor}
\usage{
extract_generic.two(page, pattern, interesting.hit)
}
\arguments{
\item{page}{page strings from document}

\item{pattern}{pattern to look for}

\item{interesting.hit}{which one to choose}
}
\value{
a number
}
\description{
Generic extractor
}
