% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapper.R
\name{extract_generic2}
\alias{extract_generic2}
\title{Generic extractor}
\usage{
extract_generic2(page, pattern.name, interesting.hit, add.me = 0, last = FALSE)
}
\arguments{
\item{page}{page strings from document}

\item{interesting.hit}{which one to choose}

\item{pattern}{pattern to look for}
}
\value{
a number
}
\description{
Generic extractor
}
