% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapper.R
\name{extract_generic_2020_12_22}
\alias{extract_generic_2020_12_22}
\title{Generic extractor}
\usage{
extract_generic_2020_12_22(page, pattern.name, pattern.list.override = NULL)
}
\arguments{
\item{page}{page strings from document}

\item{pattern}{pattern to look for}

\item{interesting.hit}{which one to choose}
}
\value{
a number
}
\description{
Generic extractor
}
\examples{
extract_generic_2020_12_22(page1, 'cases')
extract_generic_2020_12_22(page1, 'deaths')
extract_generic_2020_12_22(page1, 'recoveries')
extract_generic_2020_12_22(page1, 'hospitalized')
extract_generic_2020_12_22(page1, 'icu')
}
