% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapper.R
\name{extract_info}
\alias{extract_info}
\title{Extract information from latest report of DGS}
\usage{
extract_info(only.date = NULL, index = 1)
}
\arguments{
\item{only.date}{filter out if date isn't the same (NULL for keep whatever report it is)}

\item{index}{index of report to extract (1 is the latest)}
}
\value{
list with information that can be extracted
}
\description{
Extract information from latest report of DGS
}
\examples{
extract_info(index = 4) # keep only if it's from today
}
