% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapper.R
\name{extract_info2}
\alias{extract_info2}
\title{Extract information from latest report of DGS from 2020/08/17}
\usage{
extract_info2(only.date = NULL, index = 1)
}
\arguments{
\item{only.date}{filter out if date isn't the same (NULL for keep whatever report it is)}

\item{index}{index of report to extract (1 is the latest)}
}
\value{
list with information that can be extracted
}
\description{
Extract information from latest report of DGS from 2020/08/17
}
\examples{
extract_info2(index = 11) # keep only if it's from today
index <- 15
extract_info2(index = index) # keep only if it's from today
}
