% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_scrapper.R
\name{extract_recoveries}
\alias{extract_recoveries}
\title{Extract number of recoveries from report}
\usage{
extract_recoveries(page)
}
\arguments{
\item{page}{page strings from document}
}
\value{
a number
}
\description{
Extract number of recoveries from report
}
