% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_2020_12_22_scrapper.R
\name{extract_recoveries2}
\alias{extract_recoveries2}
\title{Extract number of recoveries from report (from 2020/08/17)}
\usage{
extract_recoveries2(page)
}
\arguments{
\item{page}{page strings from document}
}
\value{
a number
}
\description{
Extract number of recoveries from report (from 2020/08/17)
}
