% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapper.R
\name{extract_tests}
\alias{extract_tests}
\title{Extract number of tests from report}
\usage{
extract_tests(page)
}
\arguments{
\item{page}{page strings from document}
}
\value{
a number
}
\description{
Extract number of tests from report
}
