% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get.age.data}
\alias{get.age.data}
\title{Get Age data for demographics plots}
\usage{
get.age.data(input.data)
}
\arguments{
\item{input.data}{a dgs.pt tibble from this packaged}
}
\value{
a demographics ready tibble
}
\description{
Get Age data for demographics plots
}
\examples{
get.age.data(covid19.pt.data::dgs.pt)
}
