% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get.age.data.with.labels}
\alias{get.age.data.with.labels}
\title{Get age data tibble with readable labels}
\usage{
get.age.data.with.labels(input.data, date.ix)
}
\arguments{
\item{input.data}{and output of get.age.data()}

\item{date.ix}{date to filter through}
}
\value{
tibble with readable labels
}
\description{
Get age data tibble with readable labels
}
