% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get.age.new.data}
\alias{get.age.new.data}
\title{Title}
\usage{
get.age.new.data(input.data, input.data.with.labels, date.ix)
}
\arguments{
\item{input.data}{output of get.age.data}

\item{input.data.with.labels}{output of get.age.data.with.labels}

\item{date.ix}{date to filter through}
}
\value{

}
\description{
Title
}
