% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapper.R
\name{merge_eu.cdc}
\alias{merge_eu.cdc}
\title{Merge DGS and EU CDC data (that is downloaded inside function)}
\usage{
merge_eu.cdc(dgs.pt.new)
}
\arguments{
\item{dgs.pt.new}{dgs data}
}
\value{
consolidated data for Portugal
}
\description{
Merge DGS and EU CDC data (that is downloaded inside function)
}
