% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapper.R
\name{my.pattern_2020_12_22}
\alias{my.pattern_2020_12_22}
\title{Patterns to look at for individual data}
\usage{
my.pattern_2020_12_22(index)
}
\arguments{
\item{index}{name of pattern to extract}
}
\value{
a number
}
\description{
Patterns to look at for individual data
}
