% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discord.R
\name{send.discord.msg}
\alias{send.discord.msg}
\title{Send message with differences in data}
\usage{
send.discord.msg(new.dat, old.dat)
}
\arguments{
\item{new.dat}{new data}

\item{old.dat}{old data, from previous day}
}
\description{
Send message with differences in data
}
\examples{
dat1 <- dgs.pt
dat2 <- dat1
dat2[1, 3] <- 900000
send.discord.msg(dat2, dat1)
}
