% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_2020_12_22_scrapper.R
\name{extract_generic2}
\alias{extract_generic2}
\title{Generic extractor}
\usage{
extract_generic2(
  page,
  pattern.name,
  interesting.hit.force = NULL,
  add.me.force = NULL,
  last.force = NULL,
  allow.recur.force = NULL
)
}
\arguments{
\item{page}{page strings from document}

\item{pattern}{pattern to look for}

\item{interesting.hit}{which one to choose}
}
\value{
a number
}
\description{
Generic extractor
}
